Add schema-specific config from model config to schema config.
Previously the definition of the root schema file used by a model was contained entirely in the model config. Now the schema files root directory in which that file is contained, is defined in the schema config (as it will be the same for all models using that schema config).