Add fill color contrast for new rows from previous Book conversions.
When creating a new Book based on a previous Book, if a row is new (i.e. not found in based-on Book), a specific "ConversionNewRowStyle" is assigned to make the row easily noticeable by the user. However, when a Book created in this same way is subsequently used as the basis for creating a new Book, if the "ConversionNewRowStyle" remains in the based-on Book, and then new rows exist in the new Book, there would be no way for the users to tell them apart. As such, a new "ConversionPrevNewRowStyle" has been added. If a cell is copied over from the based-on Book, and it has the same fill color as the "ConversionNewRowStyle", then the style will be changed to "ConversionPrevNewRowStyle". In this way the "new row" style can be "downgraded" to still provide a means for the user to know it was, at one time, "new", but is not a new cell in the latest Book.